% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{browseFlowHist}
\alias{browseFlowHist}
\title{browseFlowHist}
\usage{
browseFlowHist(flowList)
}
\arguments{
\item{flowList}{a list of \code{FlowHist} objects}
}
\value{
Returns the list of \code{FlowHist} objects, updated by any
  changes made in the GUI.
}
\description{
Visually assess histogram fits
}
\details{
Visually assess histogram fits
}
\author{
Tyler Smith
}

