% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowHist.R
\name{exportFlowHist}
\alias{exportFlowHist}
\title{exportFlowHist}
\usage{
exportFlowHist(fh, file = NULL)
}
\arguments{
\item{fh}{a flowHist object, or a list of flowHist objects.}

\item{file}{character, the name of the file to save dat to}
}
\value{
a data frame
}
\description{
Extract analysis results from a flowHist object
}
\details{
A convenience function for extracting the results of the NLS
  curve-fitting analysis on a flowHist object.

If \code{fh} is a single flowHist object, a data.frame with a single
row is returned. If \code{fh} is a list of \code{flowHist} objects, a
row for each object will be added to the data.frame.

If a file name is provided, the data will be saved to that file.
}
\author{
Tyler Smith
}

