% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhDoRCS}
\alias{fhDoRCS}
\title{Calculcate the Residual Chi-Square for a \code{\link{FlowHist}} object}
\usage{
fhDoRCS(fh)
}
\arguments{
\item{fh}{a \code{\link{FlowHist}} object}
}
\value{
The updated \code{\link{FlowHist}} object.
}
\description{
Calculcate the Residual Chi-Square for a \code{\link{FlowHist}} object
}
\seealso{
\code{\link{fhDoCV}}, \code{\link{fhDoNLS}},
  \code{\link{fhDoCounts}}
}
\author{
Tyler Smith
}
\keyword{internal}
