% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{fhPeakPlot}
\alias{fhPeakPlot}
\title{fhPeakPlot}
\usage{
fhPeakPlot(fh, window, smooth = window/2)
}
\arguments{
\item{fh}{a \code{flowHist} object}

\item{window}{an integer, the width of the moving window to use in
identifying local maxima via \code{caTools::runmax}}

\item{smooth}{an integer, the width of the moving window to use in
removing noise via \code{caTools::runmean}}
}
\value{
None, used for plotting
}
\description{
Plot detected peaks along with the raw and smoothed data.
}
\details{
For examining/debugging the peak-finding algorithm. Allows users to see
the raw data, the smoothed data, and the peaks that are detected based
on the smoothed parameters.

A convenience for examining how different values of \code{window} and
\code{smooth} work in peak detection. Not exported for regular use by
users.
}
\author{
Tyler Smith
}
\seealso{
\code{findPeaks}
}

