% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowHist.R
\name{flowHist}
\alias{buildHist}
\alias{flowHist}
\alias{histBatch}
\title{flowHist}
\usage{
flowHist(FCS = NULL, FILE = NULL, channel, bins = 256, window = 20,
  smooth = 20, pick = FALSE)

histBatch(files, channel, bins = 256, verbose = TRUE)

buildHist(FCS, channel, bins)
}
\arguments{
\item{FCS}{a \code{flowFrame} object, as created by the \code{flowCore::read.FCS} function}

\item{FILE}{the path to an FCS file}

\item{channel}{the name of the channel containing the intensity data
for the histogram}

\item{bins}{an integer specifying the number of bins to use in building
the histogram}

\item{window}{the width of the moving window used to identify local
maxima for peak detection via \code{caTools:runmax}}

\item{smooth}{the width of the moving window used to reduce noise in
the histogram via \code{caTools::runmean}}

\item{pick}{boolean; if TRUE, the user will be prompted to select peaks
to use for starting values. Otherwise (the default), starting values
will be detected automatically.}

\item{files}{a character vector, FCS file names}

\item{verbose}{boolean; if TRUE, \code{histBatch} will list files as it
processes them.}
}
\value{
A \code{flowHist} object. Initially, it has the following
  slots:

\itemize{

\item data: a dataframe containing the histogram data. Column 'x' is
the index position, and column 'intensity' contains the fluorescence
intensity at that position.

\item file: the filename of the source data

\item peaks: a matrix containing the location of the G1 peaks, as
determined by the user or the automated peak-finding routine.

\item comps: a list of the individual model components included in the
model.

\item model: the complete model used in the NLS regression

\item init: the inital parameter estimates used in the NLS regression

}

Additionally, after the data is analyzed by supporting functions, the
following slots will be added:

\itemize{

\item nls: the fitted nls model for the data

\item counts: the number of modelled events in each model component

\item cv: the coefficients of variation for the modelled peaks, and the
confidence interval for the ratio of the sample and co-chopped standard
G1 peaks

\item RCS: the regularized Chi-Square value for the regression. There
are some issues with the interpretation of this value. It can be quite
sensitive to values at the end of the histogram, as one example.

\code{histBatch} returns a list of \code{flowHist} objects.

}
}
\description{
Create flowHist objects from an FCS file or a flowFrame object
}
\details{
Creates a \code{flowHist} object from an FCS file, or a
\code{flowFrame} object already in memory

Starting with a \code{flowFrame} object, either read from a FCS data
file or already in memory, \code{flowHist} will:

\enumerate{
\item Extract the intensity data from channel. The actual channel to
use will depend on the original FCS file. In our case, we use
"FL3.INT.LIN". See the examples for hints on how to find the right
channel.

\item Remove the top bin, which contains off-scale readings we ignore
in the analysis.

\item aggregates the raw data into the desired number of bins, as
specified with the 'bins' argument. The default is 256, but you may
also try 128 or 512. Any integer is technically acceptable, but I
wouldn't stray from the default without a good reason.

\item Identify starting values for Gaussian model components. For
reasonably clean data, the built-in peak detection is fine. You can
evaluate this by plotting the \code{flowHist} object with the argument
\code{init = TRUE}. If it doesn't look good, you can play with the
\code{window} and \code{smooth} arguments, or pick the peaks yourself
with \code{pick = TRUE}.

\item Build the NLS model. All \code{flowHist} objects will have
components for the single cut debris model and the G1 peaks for the
sample and the internal standard. If the G1 peaks are below half (i.e.,
the are at < 128 on a 256 bin histogram), their G2 peaks will be
modelled as well. All the components are combined into a single model.

\item Initial values for the model parameters are determined, using the
function \code{flowInit}.
}
}
\examples{
\dontrun{
## To find possible values for channel:
fcs <- read.FCS("path-to-file.lmd", dataset = 1, alter.names = TRUE)
fcs
## This will present a summary like this:
##
## flowFrame object 'filename'
## with 11369 cells and 9 observables:
##             name         desc range minRange maxRange
## $P1   FS.INT.LIN   FS INT LIN  1024        0     1023
## $P2   SS.INT.LIN   SS INT LIN  1024        0     1023
## $P3         TIME         TIME  1024        0     1023
## $P4  FL3.INT.LIN  FL3 INT LIN  1024        0     1023
## $P5 FL3.PEAK.LIN FL3 PEAK LIN  1024        0     1023
## $P6  FS.PEAK.LIN  FS PEAK LIN  1024        0     1023
## $P7  SS.PEAK.LIN  SS PEAK LIN  1024        0     1023
## $P8   FS.TOF.LIN   FS TOF LIN  1024        0     1023
## $P9   SS.TOF.LIN   SS TOF LIN  1024        0     1023
## 241 keywords are stored in the 'description' slot

## possible channels to use are listed in the 'name' column.

}
}
\author{
Tyler Smith
}

