% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flowShiny}
\alias{flowShiny}
\title{flowShiny}
\usage{
flowShiny(flowList)
}
\arguments{
\item{flowList}{a list of \code{flowHist} objects}
}
\value{
Returns the list of \code{flowHist} objects, updated by any
  changes made in the GUI.
}
\description{
Visually assess histogram fits
}
\details{
Visually assess histogram fits
}
\author{
Tyler Smith
}

