% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeModel}
\alias{makeModel}
\title{Build an NLS model from a list of components}
\usage{
makeModel(components, env = parent.frame())
}
\arguments{
\item{components}{a list of model components to combine}

\item{env}{the parent frame. Not intended for use by users.}
}
\value{
a function for use in the R nonlinear regression routine.
}
\description{
Build an NLS model from a list of components
}
\author{
Tyler Smith
}

