% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowHist.R
\name{plot.flowHist}
\alias{plot.flowHist}
\title{plot.flowHist}
\usage{
\method{plot}{flowHist}(x, init = FALSE, nls = TRUE, comps = TRUE, ...)
}
\arguments{
\item{x}{a flowHist object}

\item{init}{boolean; if TRUE, plot the regression model using the
initial parameter estimates over the raw data.}

\item{nls}{boolean; if TRUE, plot the fitted regression model over the
raw data (i.e., using the final parameter values)}

\item{comps}{boolean; if TRUE, plot the individual model components
over the raw data.}

\item{...}{additional arguments passed on to plot()}
}
\value{
Not applicable
}
\description{
Plot histograms for flowHist objects
}
\details{
.. content for details ..
}
\author{
Tyler Smith
}

