% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowHist.R
\name{plotFH}
\alias{plotFH}
\title{Plot the raw data for a flowHist object}
\usage{
plotFH(self, ...)
}
\arguments{
\item{self}{a flowHist object}

\item{...}{additional parameters passed to \code{plot}}
}
\value{
Not applicable, used for plotting
}
\description{
Creates a simple plot of the raw histogram data. Used as a utility for
other plotting functions, and perhaps useful for users who wish to
create their own plotting routines.
}
\author{
Tyler Smith
}

