% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{setBins}
\alias{setBins}
\title{setBins}
\usage{
setBins(fh, bins = 256)
}
\arguments{
\item{fh}{a \code{FlowHist} object}

\item{bins}{integer, the number of bins to use in aggregating FCS data}
}
\value{
a \code{FlowHist} object, with the \code{bins} slot set to
  \code{bins}, and the corresonding binned data stored in a matrix in
  the \code{histData} slot. Any previous analysis slots are removed:
  \code{peaks, comps, model, init, nls, counts, CV, RCS}.
}
\description{
(Re-) set the bins for a FlowHist object
}
\details{
This function sets (or resets) the number of bins to use in aggregating
FCS data into a histogram, and generates the corresponding data matrix.

The \code{histData} matrix also contains the \code{SCvals} column. This
is used to calculate the single-cut debris component in the NLS model.
}
\author{
Tyler Smith
}

