% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{singleCut}
\alias{singleCut}
\alias{singleCutBase}
\title{Single-cut debris model}
\usage{
singleCutBase(SCa, intensity, xx)

singleCut(SCa, intensity, xx)
}
\arguments{
\item{SCa}{a numeric value, the single-cut amplitude parameter}

\item{intensity}{a numeric vector, the histogram intensity in each channel}

\item{xx}{an integer vector, the ordered channels corresponding to the
values in `intensity'.}
}
\value{

}
\description{
Models debris using the single-cut model described by Bagwell et al.
(1991).
}
\details{
The model is:
S(x) = SCa ⱼ₌ₓ₊₁∑ⁿ ³√j YⱼPₛ(j, x)

x is the histogram channel that we're estimating the debris value for
SCa is the amplitude parameter
Yⱼ is the histogram intensity for channel j.

where Pₛ(j, x) is the probability of a nuclei from channel j falling
into channel x when cut. That is, for j > x, the probability that
fragmenting a nuclei from channel j with a single cut will produce a
fragment of size x. This probability is calculated as:

Pₛ(j, x) = 2 / (πj √(x/j)(1 - x/j)

This model involves a recursive calculation, since the fitted value for
channel x depends not just on the intensity for channel x, but also the
intensities at all channels > x. Consequently, this is coded with an
internal loop, and then vectorized to produce a well-behaved function
that we can use with the standard nls toolchain.
}
\author{
Tyler Smith
}
\references{
Bagwell, C. B., Mayo, S. W., Whetstone, S. D., Hitchcox, S.
  A., Baker, D. R., Herbert, D. J., Weaver, D. L., Jones, M. A. and
  Lovett, E. J. (1991), DNA histogram debris theory and compensation.
  Cytometry, 12: 107–118. doi: 10.1002/cyto.990120203
}

