% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhAnalyze}
\alias{fhAnalyze}
\alias{fhCV}
\alias{fhCount}
\alias{fhNLS}
\alias{fhRCS}
\title{fhAnalyze}
\usage{
fhAnalyze(fh)

fhRCS(fh)

fhNLS(fh)

fhCount(fh, lower = 0, upper = 256, subdivisions = 1000)

fhCV(fh)
}
\arguments{
\item{fh}{a flowHist object}
}
\value{
a flowHist object with the analysis (nls, counts, cv, RCS)
  slots filled.
}
\description{
Complete non-linear regression analysis of flowHist histogram data
}
\details{
Completes the NLS analysis, and calculates the modelled events and CVs
for the result.
}
\author{
Tyler Smith
}
\seealso{
\code{\link{flowHist}}
}

